# specs/001-prd-app-only/contracts/api.yaml
openapi: 3.0.3
info:
  title: Yabaii - Japanese Price Comparison App API
  description: RESTful API for Japanese price comparison mobile application
  version: 1.0.0
  contact:
    name: API Team
    url: https://yabaii.day
  license:
    name: MIT
servers:
  - url: https://api.yabaii.day/v1
    description: Production
  - url: https://staging-api.yabaii.day/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /items/{spuId}/offers:
    get:
      tags: [Products]
      summary: Get product offers from all platforms
      operationId: getProductOffers
      parameters:
        - name: spuId
          in: path
          required: true
          schema:
            type: string
        - name: platform
          in: query
          schema:
            type: string
        - name: condition
          in: query
          schema:
            type: string
            enum: [new, used, refurbished]
      responses:
        '200':
          description: Product offers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductOffersResponse'

  # Price endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    ProductOffer:
      type: object
      required:
        - id
        - productId
        - platform
        - price
        - availability
        - url
      properties:
        id:
          type: string
        productId:
          type: string
        platform:
          $ref: '#/components/schemas/Platform'
        title:
          type: string
        price:
          $ref: '#/components/schemas/Price'
        availability:
          $ref: '#/components/schemas/Availability'
        seller:
          $ref: '#/components/schemas/Seller'
        shipping:
          $ref: '#/components/schemas/ShippingInfo'
        condition:
          type: string
          enum: [new, used, refurbished]
        url:
          type: string
        images:
          type: array
          items:
            $ref: '#/components/schemas/ProductImage'
        reviews:
          $ref: '#/components/schemas/ReviewSummary'

    # Response schemas
...
    Price:
      type: object
      required:
        - amount
        - currency
      properties:
        amount:
          type: number
        currency:
          type: string
        originalPrice:
          type: number
        discountPercentage:
          type: number
    Availability:
      type: object
      required:
        - inStock
      properties:
        inStock:
          type: boolean
        quantity:
          type: integer
        estimatedDelivery:
          type: string
        stockStatus:
          type: string
          enum: [in_stock, low_stock, out_of_stock, pre_order]
    Platform:
      type: object
      required:
        - id
        - name
        - domain
      properties:
        id:
          type: string
        name:
          type: string
        domain:
          type: string
        icon:
          type: string
...
    ProductImage:
      type: object
      required:
        - url
        - width
        - height
      properties:
        url:
          type: string
        alt:
          type: object
          additionalProperties:
            type: string
        width:
          type: integer
        height:
          type: integer
        order:
          type: integer
...
    ReviewSummary:
      type: object
      properties:
        averageRating:
          type: number
        totalReviews:
          type: integer
        aiSummary:
          $ref: '#/components/schemas/AISummary'
    AISummary:
      type: object
      properties:
        summary:
          type: string
        pros:
          type: array
          items:
            type: string
        cons:
          type: array
          items:
            type: string
        confidence:
          type: number
...
    Seller:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        url:
          type: string
    ShippingInfo:
      type: object
      properties:
        free:
          type: boolean
        cost:
          type: number
        estimatedDays:
          type: integer
        methods:
          type: array
          items:
            type: string
...
    ProductOffersResponse:
      type: object
      properties:
        offers:
          type: array
          items:
            $ref: '#/components/schemas/ProductOffer'
...